appbackend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
stripe>=7.0
python-dotenv>=1.0
